Refactor coercion of vector elements from BUILDARGS back to 'coerce'
When I refactored the declaration of the 'elements' attribute from
concrete types to the Vector role, I implemented the coercion of the
vector elements to the form expected by the concrete type (e.g.,
flattenning and casting to true/false for Logical) in a BUILDARGS
'around' advice. A cleaner, and arguably the intended, way to do this
in MOO is to overload the attribute by appending a 'coerce' method, as
done by this commit.