Use scalars in Perl rep of list for atomic vectors of length 1
The idea is that in R, `1:3`, and `list(1, 2, 3)` can often be used
interchangeably, even though the list is really composed of three
integer vectors, each of length one. Now, both will have native Perl
representation of `[1, 2, 3]`.

This only applies to elements that are atomic vectors. An element of
type list will always be represented as an array reference:

`list(list(1), list(2), list(3))` -> `[ [ 1 ], [ 2 ], [ 3 ] ]`