Uncompress R data files from the loaded buffer instead of re-reading
Instead of `seek`-ing back to the beginning of the RData/RDS file when
I detect the magic marker for their being compressed, I can just have
IO::Uncompress work directly on the buffer into which the data was
loaded. This a) avoids accessing the file twice; and b) eliminates the
need to use `seek`. Not using `seek` is particularly important because
I can't call `$fh->seek` on Perl 5.10 when $fh is a file handle, and
because, strictly speaking, the constraint for $fh is that it be a
file handle, and not necessarily seekable.