Add a subclass of REXP::Closure with looser equality check for QAP tests (closes #32)
QAP doesn't include a serialization of a closure environment, which
means that in my unit tests dealing with Rserve and QAP parsing, I
have duplicate code section that cleans up the expected value of
closures. This change creates a subclass of Closure that is used only
in testing, in which the equal method is overriden to allow equality
when the compared object's "environment" is undef if everything else
matches.