Allow XT_SYM with zero length
The Symbol class allows zero length, and the `count` combinator handles
zero-count properly, so there is no reason to `die` when QAP parsing
encounters an XT_SYM with zero length.

In practice, this was never an issue because QAP pads messages to be
of length that is a multiple of 4, so a message containing just an
empty symbol is represented as '0a 08 00 00 13 04 00 00 00 00 00 00'.
But I think it may be possible to construct an object that contains an
empty symbol somewhere that ends up represented as '13 00 00 00', and
with this change we can now parse it.